Experimentally suggest that the private types used for Dolby/AC-3 might be being used for such (otherwise, it can be very unobvious that this is so).
--HG--
extra : convert_revision : svn%3Aeff31bef-be4a-0410-a8fe-e47997df2690/trunk%4016

+++ ts2es.c
@@ -267,7 +267,7 @@ static int extract_pid_packets(TS_reader_p  tsreader,
         if (written != 1)
         {
           fprintf(stderr,"### Error writing TS packet - units written = %d\n",
-                  written);
+                  (int)written);
           return 1;
         }
       }
